+++ src/algo/mod.rs
@@ -4,8 +4,8 @@
 
 pub mod bellman_ford_moore;
 pub mod bfs;
+pub mod bfs_tree;
 pub mod dfs;
 pub mod dijkstra;
 pub mod fixture;
 pub mod floyd_warshall;
-pub mod predecessor;